cli: clean cmd migrated to docopts.
This also involved moving `load_config` to the `yaml` module.
